% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_initial_guess.R
\name{beta_initial_guess}
\alias{beta_initial_guess}
\title{Beta Initial Guess}
\usage{
beta_initial_guess(x, y, data)
}
\arguments{
\item{x}{A \code{character vector} containing the names of all features
we will train the binary classification on.}

\item{y}{A \code{character} containing the name of the classified feature
(which we will fit the model on).}

\item{data}{A \code{data frame} containing the data we will train the binary
classification on.}
}
\value{
A p x 1 \code{vector} containing the initial beta coefficient values,
where p is the number of features given plus 1 for the intercept.
}
\description{
Determine Initial Beta Coefficients for optimization using the
least-squares formula.
}
\examples{
beta_initial_guess(x = c("age", "workclass",
       "hours.per.week"), y = "sex", data = adult)
}
\author{
Emily Knight
}
