% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{bootstrapping}
\alias{bootstrapping}
\title{Bootstrapping}
\usage{
bootstrapping(x, y, data, alpha = 0.05, B = 20)
}
\arguments{
\item{x}{A \code{character vector} containing the names of all features
we will train the binary classification on.}

\item{y}{A \code{character} containing the name of the classified feature
(which we will fit the model on).}

\item{data}{A \code{data frame} containing the data we will train the binary
classification on.}

\item{alpha}{A \code{numeric} representing the significance level for the 1 - alpha
confidence intervals. Default set to 0.05.}

\item{B}{A \code{integer} representing the number of of bootstraps to run. Default
set to 20.}
}
\value{
A \code{list} containing the following attributes, where p is the number of
provided features to predict on (including the intercept) and B is the number of bootstraps:
\describe{
     \item{Coefficients}{A list containing the following matrices:
         \describe{
             \item{Beta_Bootstrap_Estimates}{A B x p \code{matrix} containing the
             optimized beta coefficients for each bootstrap iteration. (The bth row
             contains the optimized beta for the bth bootstrap.)}
             \item{Beta_Boostrap_Initial}{A B x p \code{matrix} containing the initial
             beta values for each bootstrap iteration, obtained from the least-squares
             formula. (The bth row contains the initial beta for the bth bootstrap.)}
         }
     }
     \item{ConfidenceIntervals}{A 3 x p matrix containing the confidence intervals for
     beta for each feature. The first row, Lower, contains the lower estimate; the second row,
     Estimate, contains the mean from the bootstrap procedure; the third row, Upper, contains
     the upper estimate.}
}
}
\description{
Given input data, confidence level, and number of runs, run the bootstrap
procedure to obtain confidence intervals for beta estimates.
}
\examples{
bootstrapping(x = c("age", "workclass", "hours.per.week"),
     y = "sex", data = adult, alpha = 0.1, B = 100)
}
\author{
Emily Knight
}
