% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_metrics.R
\name{confusion_metrics}
\alias{confusion_metrics}
\title{Confusion Metrics}
\usage{
confusion_metrics(beta, x, y)
}
\arguments{
\item{beta}{A \code{vector} of size p x 1, where p is the number of features in the data
(including the intercept), representing the coefficients we will use to predict with logistic
regression.}

\item{x}{A \code{matrix} of size n x p, where n is the number of observations of the data
and $p$ is the number of features (including the intercept).}

\item{y}{A \code{vector} of size n x 1 representing the classification of each of the n
observations, 0 or 1.}
}
\value{
A \code{list} containing the following attributes:
\describe{
\item{ConfusionMatrix}{A \code{matrix} of integers that represent how many of each classification
logistic regression classified correctly and incorrectly using the provided beta vector}
\item{Metrics}{A \code{list} of six numeric values which are, in order:
     \describe{
          \item{Prevalence}{}
          \item{Accuracy}{}
          \item{Sensitivity}{}
          \item{Specificity}{}
          \item{FalseDiscoveryRate}{}
          \item{DiagnosticOddsRatio}{}
     }
     }
   }
}
\description{
Given coefficient vector beta, feature data, and classifications, output the
predictions of logistic regression with the correct classifications in a confusion matrix.
Also output prevalence, accuracy, sensitivity, specificity, false discovery rate, and
diagnostic odds ratio.
}
\author{
Shuqi Du
}
