% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_ci.R
\name{graph_ci}
\alias{graph_ci}
\title{Graph Bootstrap Boxplots}
\usage{
graph_ci()
}
\arguments{
\item{x}{A \code{character vector} containing the names of all features
we will train the binary classification on.}

\item{y}{A \code{character} containing the name of the classified feature
(which we will fit the model on).}

\item{data}{A \code{data frame} containing the data we will train the binary
classification on.}

\item{alpha}{A \code{numeric} representing the significance level for the 1 - alpha
confidence intervals. Default set to 0.05.}

\item{B}{A \code{integer} representing the number of of bootstraps to run. Default
set to 20.}
}
\value{
A graph with p boxplots, where p is the number of features you ran the bootstrap
procedure on plus the intercept. The y axis shows the value for the beta feature, and the
x axis has the p boxplots.
}
\description{
Run the bootstrapping procedure with the bootstrapping function
and produce boxplots.
}
\examples{
graph_ci(x = c("age", "workclass", "hours.per.week"),
     y = "sex", data = adult, alpha = 0.1, B = 100)
}
\author{
Emily Knight
}
