% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphci.R
\name{graphci}
\alias{graphci}
\title{Graph Bootstrap Boxplots}
\usage{
graphci(x, y, data, alpha = 0.05, B = 20)
}
\arguments{
\item{x}{A \code{character vector} containing the names of all features
we will train the binary classification on.}

\item{y}{A \code{character} containing the name of the classified feature
(which we will fit the model on).}

\item{data}{A \code{data frame} containing the data we will train the binary
classification on.}

\item{alpha}{A \code{numeric} representing the significance level for the 1 - alpha
confidence intervals. Default set to 0.05.}

\item{B}{A \code{integer} representing the number of bootstraps to run. Default
set to 20.}
}
\value{
Produces a graph with p boxplots, where p is the number of features you ran the bootstrap
procedure on plus the intercept. The y axis shows the value for the beta feature, and the
x axis has the p boxplots. Also returns the output of the bootstrapping procedure (see
help documentation for bootstrapping function for details.)
}
\description{
Run the bootstrapping procedure with the bootstrapping function
and produce boxplots.
}
\author{
Emily Knight
}
